requirements.txt
google-genai # For Google Generative AI
pillow # For image processing
pyautogui # For GUI automation
langchain # For building applications with LLMs
langchain-community # For community-driven LangChain extensions
langchain-experimental # For experimental features in LangChain
langgraph # For building graph-based applications with LLMs
langchain-google-genai # LangChain integration for Google Generative AI
scikit-image # For image processing and computer vision tasks